+++ database/factories/TokenFactory.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+|--------------------------------------------------------------------------
+| Token Factory
+|--------------------------------------------------------------------------
+*/
+
+$factory->define(\JobApis\JobsToMail\Models\Token::class, function (Faker\Generator $faker) {
+    return [
+        'user_id' => null,
+        'type' => 'confirm',
+    ];
+});