+++ database/seeds/TestingDatabaseSeeder.php
@@ -4,7 +4,6 @@ use Faker\Factory as Faker;
 use Illuminate\Database\Seeder;
 use JobApis\JobsToMail\Models\User;
 use JobApis\JobsToMail\Models\Token;
-use Illuminate\Support\Facades\DB;
 
 class TestingDatabaseSeeder extends Seeder
 {
@@ -18,56 +17,36 @@ class TestingDatabaseSeeder extends Seeder
 
     private function createActiveUsers($num = 10)
     {
-        foreach(range(1, $num) as $index)
-        {
-            $user = User::create([
-                'email' => $this->faker->email(),
-                'keyword' => $this->faker->word(),
-                'location' => $this->faker->word().', '.$this->faker->word(),
-                'confirmed_at' => $this->faker->dateTimeThisYear(),
-            ]);
-            Token::create([
-                'user_id' => $user['id'],
-                'type' => 'confirm',
-            ]);
-        }
+        return factory(User::class, $num)
+            ->states('active')
+            ->create()
+            ->each(function(User $user) {
+                $user->tokens()->save(
+                    factory(Token::class)->make()
+                );
+            });
     }
 
     private function createDeletedUsers($num = 10)
     {
-        foreach(range(1, $num) as $index)
-        {
-            $user = User::create([
-                'email' => $this->faker->email(),
-                'keyword' => $this->faker->word(),
-                'location' => $this->faker->word().', '.$this->faker->word(),
-                'confirmed_at' => $this->faker->dateTimeThisYear(),
-                'deleted_at' => $this->faker->dateTimeThisYear(),
-            ]);
-            Token::create([
-                'user_id' => $user['id'],
-                'type' => 'confirm',
-            ]);
-            Token::create([
-                'user_id' => $user['id'],
-                'type' => 'unsubscribe',
-            ]);
-        }
+        return factory(User::class, $num)
+            ->states('active', 'deleted')
+            ->create()
+            ->each(function(User $user) {
+                $user->tokens()->save(
+                    factory(Token::class)->make()
+                );
+            });
     }
 
     private function createUnconfirmedUsers($num = 10)
     {
-        foreach(range(1, $num) as $index)
-        {
-            $user = User::create([
-                'email' => $this->faker->email(),
-                'keyword' => $this->faker->word(),
-                'location' => $this->faker->word().', '.$this->faker->word(),
-            ]);
-            Token::create([
-                'user_id' => $user['id'],
-                'type' => 'confirm',
-            ]);
-        }
+        return factory(User::class, $num)
+            ->create()
+            ->each(function(User $user) {
+                $user->tokens()->save(
+                    factory(Token::class)->make()
+                );
+            });
     }
 }