+++ changelog.md
@@ -1,6 +1,19 @@
 # Changelog
 All Notable changes to `jobs-to-mail` will be documented in this file.
 
+## 0.4.0 - 2016-10-14
+
+### Added
+- Support for [Ziprecruiter](https://github.com/jobapis/jobs-ziprecruiter) job board.
+- Automatically run migrations after composer install.
+- Logging errors from collections.
+- Model factories for database seed operations.
+- Command line argument to run collection job for single email.
+
+## Fixed
+- Bug in Careerbuilder API via Jobs-Multi upgrade.
+- Removing HTML highlighting characters from Juju job results.
+
 ## 0.3.0 - 2016-10-06
 
 ### Added